Tier-3 audit fixes: harden .vtri reader + vendor tdc decoder hardening
Crafted or corrupt sidecar/index files must be rejected cleanly, never crash.
Regression tests in test-audit-tier3.R (a segfault would take down the test
process, so "returns or errors" is the assertion). Full suite green (3234).

.vtri reader (src/vtri.c):
* Validate n_entries / n_slots against the actual file size before allocating,
  written division-first so the size arithmetic cannot overflow. Previously a
  crafted n_entries (e.g. 2^61) overflowed (size_t)ne * 8 to a small/zero malloc
  and the fill loop wrote far past it -- a controllable heap-write overflow.
* Reject a v2 composite header with zero columns (was an OOB read of col
  index 0), and require n_slots >= 1 (n_slots == 0 made the probe mask -1).
* probe_by_hash now bounds every chain index to [0, n_entries) and caps the
  traversal at n_entries steps, so a crafted head/next link cannot read out of
  range or loop forever.

Vendored tdc bumped to 294d124 (gcol33/tdc), which adds the matching LZ
sequence-cursor / back-reference bounds and the DICT_1D 64-bit offset-table
sizing fix. Round-trip unchanged (compression/codec-tdc/vtr1-tdc/dict-defer
green).

+++ src/tdc/src/model/dict.c
@@ -327,18 +327,22 @@ static tdc_status dict1d_validate_side_header(const uint8_t *side_meta,
     memcpy(&dict_count, side_meta + 0, 4u);
     memcpy(&dict_total, side_meta + 4, 4u);
 
-    size_t want_side = (size_t)8u
-                     + (size_t)(dict_count + 1u) * 4u
-                     + (size_t)dict_total;
-    if (side_size != want_side) return TDC_E_CORRUPT;
+    /* 64-bit so (dict_count + 1) * 4 cannot wrap to 0 when dict_count is
+     * UINT32_MAX; an oversized count then fails this equality instead of
+     * zeroing the table size and leaving the loop below unterminated. */
+    uint64_t want_side = (uint64_t)8u
+                       + ((uint64_t)dict_count + 1u) * 4u
+                       + (uint64_t)dict_total;
+    if ((uint64_t)side_size != want_side) return TDC_E_CORRUPT;
 
     const uint8_t *off_p = side_meta + 8u;
 
     /* Verify the offsets table is internally consistent and matches the
      * declared total. Catches torn/corrupt records before we dereference
-     * out-of-range offsets in the inner loop. */
+     * out-of-range offsets in the inner loop. The counter is 64-bit so
+     * d <= dict_count cannot wrap near UINT32_MAX. */
     uint32_t prev_off = 0u;
-    for (uint32_t d = 0; d <= dict_count; ++d) {
+    for (uint64_t d = 0; d <= (uint64_t)dict_count; ++d) {
         uint32_t cur;
         memcpy(&cur, off_p + (size_t)d * 4u, 4u);
         if (cur < prev_off)   return TDC_E_CORRUPT;
